Code cleanup!
- Refactored module and function names
- Changed constants from functions to macros
- Consolidated application and supervisor modules
- Removed compiled binaries
- Fixed all warnings for Elixir 1.4
