+++ mix.exs
@@ -4,7 +4,7 @@ defmodule Satellite.Mixfile do
   def project do
     [app: :satellite,
      version: "0.1.0",
-     elixir: "~> 1.3",
+     elixir: "~> 1.4",
      build_embedded: Mix.env == :prod,
      start_permanent: Mix.env == :prod,
      deps: deps()]
